+++ api/grabbers/models.py
@@ -20,7 +20,7 @@ class Grabber(models.Model):
     date_to = models.DateField(blank=True, null=True)
     start_date = models.DateTimeField(blank=False, null=False)
     finish_date = models.DateTimeField(blank=True, null=True)
-    posts_count = models.IntegerField(blank=True, null=True)
+    posts_count = models.BigIntegerField(blank=True, null=True)
 
     def __str__(self):
         return f'Grabber {self.pk} for group "{self.group}" by user "{self.owner.username}"'
@@ -28,13 +28,13 @@ class Grabber(models.Model):
 
 class Post(models.Model):
     grabbers = models.ManyToManyField(Grabber, related_name='posts')
-    owner_id = models.IntegerField()
-    post_id = models.IntegerField()
+    owner_id = models.BigIntegerField()
+    post_id = models.BigIntegerField()
     date = models.DateTimeField()
     is_ad = models.BooleanField()
-    likes = models.IntegerField()
-    reposts = models.IntegerField()
-    comments = models.IntegerField()
+    likes = models.BigIntegerField()
+    reposts = models.BigIntegerField()
+    comments = models.BigIntegerField()
     has_playlist = models.BooleanField()
     has_audios = models.BooleanField()
     text = models.TextField(blank=True, null=True)
@@ -45,11 +45,11 @@ class Post(models.Model):
 
 class Playlist(models.Model):
     posts = models.ManyToManyField(Post, related_name='playlists')
-    owner_id = models.IntegerField()
-    playlist_id = models.IntegerField()
+    owner_id = models.BigIntegerField()
+    playlist_id = models.BigIntegerField()
     access_hash = models.CharField(max_length=100, blank=True, null=True)
-    listens = models.IntegerField(blank=True, null=True)
-    followers = models.IntegerField(blank=True, null=True)
+    listens = models.BigIntegerField(blank=True, null=True)
+    followers = models.BigIntegerField(blank=True, null=True)
     title = models.TextField()
     create_date = models.DateTimeField(blank=True, null=True)
     update_date = models.DateTimeField(blank=True, null=True)
@@ -61,12 +61,12 @@ class Playlist(models.Model):
 
 class Audio(models.Model):
     posts = models.ManyToManyField(Post, related_name='audios')
-    owner_id = models.IntegerField(blank=False, null=False)
-    audio_id = models.IntegerField(blank=False, null=False)
+    owner_id = models.BigIntegerField(blank=False, null=False)
+    audio_id = models.BigIntegerField(blank=False, null=False)
     artist = models.TextField(blank=False, null=False)
     title = models.TextField(blank=False, null=False)
     date = models.DateTimeField(blank=True, null=True)
-    savers_count = models.IntegerField(blank=True, null=True)
+    savers_count = models.BigIntegerField(blank=True, null=True)
     doubles = models.IntegerField(blank=True, null=True)
     source = models.TextField(blank=True, null=True)
     parsing_date = models.DateTimeField(blank=True, null=True)
