+++ api/users/models.py
@@ -7,7 +7,7 @@ class User(DjangoUser):
 
     balance = models.IntegerField(default=0)
     ads_token = models.CharField(max_length=100, blank=True, null=True)
-    user_id = models.IntegerField(blank=True, null=True)
+    user_id = models.BigIntegerField(blank=True, null=True)
     ava_url = models.TextField(default=DEFAULT_AVA_URL)
     has_token = models.BooleanField(default=False)
     can_ads = models.BooleanField(default=False)
